Add Catch2 dependency and implement initial tests for matrix operations and model training
- Introduced Catch2 as a testing framework by adding necessary files and scripts for fetching and populating the library.
- Created unit tests for matrix operations, including addition and dot product functionalities.
- Developed tests for a simple model training process, validating the structure and output of the model.
- Added JSON files for visualizations of model layers, detailing weights and biases for two dense layers.
- Included necessary CMake scripts for building and installing the Catch2 library.

+++ ML/models/cpp-modules/src/matrix.cpp
@@ -247,9 +247,9 @@ std::vector<double> solve_AX_eq_B(std::vector<std::vector<double>> A, std::vecto
 
     
     std::vector<double> X(size, 0.0);
-    for (size_t i = size - 1; i >= 0; i--) {
+    for (int i = (int)size - 1; i >= 0; i--) {
         double sum = B[i];
-        for (size_t j = i + 1; j < size; j++) {
+        for (size_t j = (size_t)i + 1; j < size; j++) {
             sum -= A[i][j] * X[j];
         }
         X[i] = sum / A[i][i];
@@ -266,10 +266,10 @@ std::vector<std::vector<double>> multiply_matrices(const std::vector<std::vector
     size_t colsB = B[0].size();
     
     std::vector<std::vector<double>> ret(rowsA, std::vector<double>(colsB, 0.0));
-    for(int i=0; i<rowsA; i++) {
-        for(int j=0; j<colsB; j++) {
+    for(size_t i=0; i<rowsA; i++) {
+        for(size_t j=0; j<colsB; j++) {
             double val = 0;
-            for(int k=0; k<colsA; k++) {
+            for(size_t k=0; k<colsA; k++) {
                 val += A[i][k] * B[k][j];
             }
             ret[i][j] = val;
